# apps/api/public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/venues": {
      "post": {
        "operationId": "VenueController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVenueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Venue"
        ]
      }
    },
    "/api/venues/{slug}": {
      "get": {
        "operationId": "VenueController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeRelations",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Venue"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dance Manager",
    "description": "Dance Manager description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
...
      "CreateVenueRequest": {
        "type": "object",
        "properties": {}
      }
    }
  }
}